refactor(rust): move DevicesChangedStream to its own module
